Mise à jour pour nouvelles versions DC
